Fix Bug And Update Feature

+++ Src/Core/Application/Common/Interfaces/IResponseWarpper.cs
@@ -19,6 +19,6 @@ public interface IResponsePageable<T>
     public string Pivot { get; set; }
     public bool HasNext { get; set; }
     public bool HasPrevious { get; set; }
-    public int CountAll { get; set; }
+    public long CountAll { get; set; }
     public int CountData { get; }
 }
