+++ Src/Core/Application/Common/models/PesponsePageable.cs
@@ -9,7 +9,7 @@ public class Pageable<T> : IResponsePageable<T>
     public virtual string Pivot { get; set; } = string.Empty;
     public virtual bool HasNext { get; set; }
     public virtual bool HasPrevious { get; set; }
-    public int CountAll { get; set; }
+    public long CountAll { get; set; }
 
     public int CountData => Collections.Count();
